requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==21.2.0
autopep8==1.5.7
backports.functools-lru-cache==1.5
cffi==1.14.6
cheroot==8.5.2
CherryPy==18.6.1
contextlib2==0.5.5
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
iniconfig==1.1.1
isort==4.3.4
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==1.20
jaraco.text==3.2.0
kiwisolver==1.3.2
lazy-object-proxy==1.3.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
more-itertools==4.3.0
numpy==1.22.0
packaging==21.0
passlib==1.7.4
Pillow==9.0.1
pluggy==1.0.0
portend==2.3
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.1.1
pyparsing==2.2.1
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.7.3
pytz==2018.5
scrypt==0.8.20
six==1.11.0
tempora==1.13
toml==0.10.2
wrapt==1.10.11
zc.lockfile==1.3.0
google-cloud-bigquery>=3.11.0,<4.0.0
